parser/requirements.txt
sqlalchemy
psycopg2-binary
loguru
requests
BeautifulSoup4
newspaper3k